+++ R/libraryAll.R
@@ -7,9 +7,11 @@
 #' @param lib.loc OPTIONAL. library store location
 #' @param quietly OPTIONAL. attach library quietly
 #' @param clear OPTIONAL. clear environment after attach
+#' @param clearPkgs Clear previous loaded packages, TRUE or FALSE
 #' @return loaded libraries and clear environment
 #' @examples
 #' \donttest{
+#' # load packages and print their versions to the console
 #' libraryAll(base) #one package
 #'
 #' libraryAll(
@@ -29,10 +31,21 @@
 #' libraryAll(
 #'   base,
 #' clear = FALSE) #do not clear console after load
+#'
+#' # clear previously loaded packages, then load r2resize and r2social
+#' libraryAll(
+#'   stats,
+#'   utils,
+#'   clearPkgs = TRUE
+#' )
 #' }
 #' @export
 
-libraryAll <- function(..., lib.loc = NULL, quietly = FALSE, clear = TRUE) {
+libraryAll <- function(..., lib.loc = NULL, quietly = FALSE, clear = TRUE, clearPkgs = FALSE) {
+  # remove previous loaded packages
+  if (clearPkgs) {
+    clearPreviouslyLoadedPkgs()
+  }
   # load quickcode if not loaded
   if("quickcode" %nin% (.packages())) library(quickcode, quietly = TRUE)
   # load user requested libraries